% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_pdf.R
\name{chrome_pdf}
\alias{chrome_pdf}
\title{Convert HTML to PDF using Chrome}
\usage{
chrome_pdf(html_path, pdf_path = NULL, temp_dir = tempfile(),
  browser_bin = NULL)
}
\arguments{
\item{html_path}{.htm/.html document to export}

\item{pdf_path}{Path to write the exported PDF document.  By default, drposter will
strip the .html extension and write a .pdf file}

\item{temp_dir}{Location of the temporary user profile directory for headless Chrome.
If you run into problems, the `decapitated` package suggests that a subdirectory of
the user's home (e.g. ~/temp_drposter) may work better}

\item{browser_bin}{Path to run Google Chrome.  If unspecified, drposter will
automatically search common installation paths for your current OS.}
}
\description{
Calls headless Chrome to export an HTML document to a PDF poster automatically
without needing to manually open up the web browser and "print" the document.
}
\details{
Use Chrome/Chromium to convert an HTML document (e.g. drposter HTML output) to PDF.
A temporary profile directory is required to launch headless Chrome as a separate
instance than any currently open windows, else the command line flags will not
take effect.

This implementation uses code from the pagedown, decapitated, and xfun packages.
}
\references{
\url{https://developers.google.com/web/updates/2017/04/headless-chrome}
\url{https://askubuntu.com/questions/35392/how-to-launch-a-new-instance-of-google-chrome-from-the-command-line}
\url{https://github.com/rstudio/pagedown/blob/master/R/chrome.R}
\url{https://www.rdocumentation.org/packages/decapitated/versions/0.3.0}
}
