% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drposter.R
\name{drposter_poster}
\alias{drposter_poster}
\title{Generate academic posters in R Markdown and CSS}
\usage{
drposter_poster(self_contained = FALSE, template = NULL,
  css = "poster.css", pandoc_args = NULL, theme = NULL,
  fill_page = FALSE, ...)
}
\arguments{
\item{self_contained}{Whether or not the poster should be packaged into a
a single self-contained file. Set to \code{FALSE} by default to save build time.}

\item{template}{Pandoc template to use for rendering. Pass \code{NULL} to use
the HTML poster template built into this package.  You can also pass a path
to a custom pandoc template that you've created.}

\item{css}{One or more CSS files defining the poster styling. Some base
styles in poster.css are copied by default into the current directory and
automatically linked in the generated HTML unless this parameter is defined.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{theme}{Currently unused.}

\item{fill_page}{Can optionally stretch the page content (by distributing
whitespace between blocks) to span the entire height assigned to the poster.
By default this is set to \code{FALSE}, which places content without extra spacing.}

\item{...}{Additional parameters to pass to html_document; otherwise ignored.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
R Markdown (pandoc) format for generating HTML/CSS posters, inspired by reveal.js
}
\details{
Use level 1 sections (`#`) to denote main divisions of content.
Assigning a `{.col-x}` class, where `x` is 1--4, will stretch the container
across the page width with `x` equal subcolumns, using CSS Grid.
The actual content goes inside of level 2 containers (`## Block title here`).
Placing multiple level 2 containers within a level 1 section will distribute
their heights to match adjacent sections in the row, allowing for multicolumn
layouts (e.g. a standard 3-column academic poster).
There are also a few convenience classes, such as formatting a QR code block.
The markdown source for the poster in \code{inst/example} perhaps provides the
best documentation by example.

The overall layout is inspired by markdown for slide presentations in reveal.js.
For additional documentation on using revealjs presentations see
\href{https://github.com/rstudio/revealjs}{https://github.com/rstudio/revealjs}.
}
\examples{
\dontrun{

library(drposter)

# simple invocation
rmarkdown::render("pres.Rmd", drposter_poster())
}

}
