% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{drposter_update}
\alias{drposter_update}
\title{Update drposter template}
\usage{
drposter_update(verbose = TRUE)
}
\arguments{
\item{verbose}{Prints backup and package directories to the console}
}
\description{
Update supporting files for drposter template in the directory drposter_files/
using a newer version from the installed package.  This function does not update your
installed version of the drposter package, just the locally cached template files.
}
\details{
Copies relevant drposter resources from the rmarkdown skeleton into the local copy
in drposter_files/.  Makes a backup by default to a directory including the current
datetime, unless drposter_files/ does not exist.  Previous versions attempted to
make the lib_dir and backup_dir configurable, but they came across compatibility issues
with the cached rmarkdown workflow and/or R's file.rename/copy commands.
}
